requirements.txt
# FastAPI
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
pydantic==1.10.7

# Machine Learning
scikit-learn==1.2.2
numpy==1.24.2
opencv-python==4.7.0.72
dlib==19.24.0

# Celery
celery==5.2.7
redis==4.5.4

# MongoDB
motor==3.1.2
pymongo==4.3.3

# HTTP Client
httpx==0.24.0
aiohttp==3.8.4

# Utilities
python-dotenv==1.0.0
Pillow==9.5.0